chore: move image merging to StageDrop analyzer
